backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
g4f==0.2.0
python-dotenv==1.0.1
cors==1.0.1